agilent: Add test to compare agilent reader to ENVI reader
Note that the ENVI reader is reading the exact same data file in the
image case (4_noimage_agg256.dat) but different files in the mosaic as
the agilent reader uses the raw mosaic tiles but envi points to the
combined .dat file.

Also XY coordinates are lost in ENVI format, so I don't test that.