Peak Fit: finally responsive (multiprocessing)
The computation of both previews and final results is done in separate
processes. Lmfit uses a scipy's optimization function that prevents
thread switching, therefore it blocked the user interface even when
executed in a separate thread.

This PR runs all lmfit's computation in separate processes. This
improves responsiveness and performance.