Removed checks for the same domain in init_interface
For me, this does not change the behaviour, because the Orange's contexts
themselves restore current values after they are reopened. I have to
admit I do not quite understand why I wrote this code in 2017 and then
copied it multiple times. Perhaps I did not use contexts then? No idea.

This is done with hopes of fixing  exceptions when variables from settings
and those available do not match. I saw bugs like this and have seen
error reports on Sentry, but I do not know how to reproduce the bug.
I do not know whether this change will fix the bugs I saw, but it seems
to remove one possibility for errors.