Fix decimals_needed_for_unique_str for cases like [1.5, 2.5]
We need to avoid possible degenerate cases like (1.5, 2.5):
the log of the difference (1.0) is 0, but rounded to zero
decimals we would get (2, 2) due to floating point rounding behavior.
Therefore, slightly increase the number of needed decimals.

Also, take one additional decimal than would be needed for
uniqueness to increase precision.